+++ addons/yat/src/enums/ECommandAdditionStatus.cs
@@ -1,0 +1,9 @@
+namespace YAT.Enums;
+
+public enum ECommandAdditionStatus
+{
+	Success,
+	UnknownCommand,
+	MissingAttribute,
+	ExistentCommand
+}